Lazy loading enabled
All pages with lots of images now have lazy loading enabled. The
"animation" was removed since it made for a bit of an odd experience one
the 2nd visit to a page. On pages with lots of images this has a nice
impact as it only loads the images that the user can see initially. As a
user scrolls it loads the next ones up. It saves a lot of extra requests
initially.