Fix season images not showing up when Emby starts.
The TvdbSeasonImageProvider was running before the TvdbSeasonImageProvider. This caused the seriesid be null on the series. (This is apparently populated as part of the metadata refresh on the series. Moving that scan before the seasons seems to fix the problem.

See the following code from TvdbSeriesImageProvider

var seriesId = series != null ? series.GetProviderId(MetadataProviders.Tvdb) : null;
if (!string.IsNullOrEmpty(seriesId) && season.IndexNumber.HasValue)