OmdbItemProvider: Use search instead of title API
In my configuration, the Omdb provider is the first in the list for
movie metadata. This was the default at the time I installed MB and I
never changed that (don't know what the current defaults are for a new
installation).
When I use the identify command for movies in the metadata editor, I
always get a single result only. This is due to the fact that the Omdb
provider used the title API with "t=moviename", while there is also a
search api with "s=moviename" which will return multiple results.

This commit modifies the OmdbItemProvider to use the search API variant
and enable returning multiple results