Auto-Organize: PathTooLongException on source file should not break auto-organize task
PathTooLongException can not only occur with long destination paths but
also with too long file names of files contained in a watch folder.
Previously this condition caused the auto-organize task to break.

With this change, we still log the exception, but auto-organize
processing will continue to handle all other files.

Conflicts:
	MediaBrowser.Server.Implementations/FileOrganization/TvFolderOrganizer.cs