Restrict metadata refresh of missing/virtual episodes
This commit can significantly improe library scan times!
(in my case from 15minutes to 50s)

This fixes the following situation:

When a user has a series in his library with lots of episodes (like
thousands), the metadata for ALL series episodes was refreshed during
each library scan. This resulted in very long scan times, even when
using the provider's cached data.

This fix skips metadata refresh for episodes that are NOT present in the
local library and are having an airdate older than 30days (no more
changes likely to occur) during library scans.

But these skipped items would still be refreshed during a full metadata
refresh.

I believe this is an acceptable balance between performance and
accuracy.