Remove 'chown' in docker container entrypoint
The entrypoint of the docker container changes the permissions of the
'/config' and '/media' directory to the environment variables PUID and PGID.

This can be used to change the ownership of any directory on the system
as long the user can run the docker command.
After owning the directory the attacker will be able to read files when
the others permission allow reading. This is mostly the default setting.

The docker container should never change any ownership of volume
directories. When the permissions of the config or media volumes are not
correct jellyfin should fail with permission denied.

Closes: https://github.com/jellyfin/jellyfin/issues/220