Add bad web build branch hotfix
I hate this quick and dirty hack but it makes no sense to port to
master. This fixes a bug whereby we'd build with the master Web branch
on releases due to never checking out the right branch. This is already
obsoleted in the master branch since #1925 already replaces this entire
process for Debuntu builds, and others should be fixed with a more
robust solution. That said, for the 10.4.z release chain, this
ultra-quick solution fixes the problem without changing much.