Added newer version of ffmpeg into the list.
Improved ffmpeg version analysis.

I had additional modules that meant it never matched the dictionary.

Only a select set of modules are now compared, and in the correct sequence as in the dictionary.

If no match is found, the nearest version to libavutil is matched and returned with a minorRevision so that the system can identify it's a custom revision based upon a version number.