Support IsLocal on a unix socket with X-Forwarded-For
Right now IsLocal will work on a unix socket because bother headers are null *unless* you set `X-Forwarded-For`. However `X-Forwarded-For` is necessary for getting the remote header. Furthermore X-Forwarded-For is necessary for websocket connections right now.

This change allows either the local or the remote address to be null. I'm not sure if there is a case where the remote address can be null and the local can't (otherwise how are we talking?) but decided to leave the check in to minimize the change.