Support `IsLocal` on a unix socket with `X-Forwarded-For`
Right now `IsLocal` will work on a unix socket because both local and remote IP addresses are null *unless* you set `X-Forwarded-For`. However `X-Forwarded-For` is necessary for getting the remote header. Furthermore `X-Forwarded-For` is necessary for websocket connections right now.

This instead first checks the remote IP, if that is null we assume a UNIX socket and return `true`. Otherwise we fall back to the IP check. This way `IsLocal` will return the results based on the `X-Forwarded-For` header if available.