Add/expose settings for video extras
Fall back to `Owner` on missing `Parent` when searching up the tree for
the library root to allow options to be found for owned items.

Fall back to `Video` when type `Trailer` not found in library options.

Add check to ensure owned videos know their owner on metadata refresh,
as `LocalAlternateVersions` could have it unset.

Add `Video` library options to allow the user to modify image
collection settings for video extras.