Update API delete item logic
Issued 7610 reported an error when using the delete item API.
It seems that there was not enough logic to handle when the user
variable returned null but the API token was authorized.

Added a new boolean that holds the API token authentication status.
If the user variable returns null from GetAuthorizationInfo then we
check to see if the API key is authenticated. If neither of these
conditions return true then Unauthorized() is returned.