Add support for "Digital Media" album splits
"Digital Media" is a common 'disk'-splitting prefix, more so with recent "digital" music releases as physical cd's/disks aren't used. 

In particular, it is part of Lidarr's {Medium Format} tag for automatic archive sorting. So it would be good to see this reflected into Jellyfin.

I'm not familiar with the code-base, or whether a ' ' character is valid within this context.