Improve TMDB matching
In some cases, mainly with n-ology, if the first episode is named only
based on the first movie name, another of the movie can be chosen from
the imdb list, even if a perfect match exists.

The current implementation takes the first result and even if tmdb
should store per relevance (close match, popularity) this is not correct

The proposed solution tries to find perfect or good match for the movie
instead of the first one.

Results for `Foo 1`
- `Foo 2`
- `Foo`
- `Foo 3`
- `The Invincible Foo`
...
Chosen Result for `Foo 1`: `Foo`

Result for `Bar 3`
- `Bar Snibar`
- `Baroid`
- `Bar`
- `Bars`
- `Bar³`
- `Bar: CoBar`
- `Bars vs Foo`
...
Chosen Result for `Bar 3`: `Bar³`

Result for `Un film Anglais`
 - `An english movie: The first episode`
 - `An english movie: The third episode`
 - `An english movie: The second episode`
 - `An english movie`
 - `After the Movie: The first episode`
...
Chosen Result for `Un film Anglais`: `An english movie: The first episode` (This was luck)

Result for `An english movie`
 - `An english movie: The second episode`
 - `An english movie: The first episode`
 - `An english movie: The third episode`
 - `An english movie`
 - `After the Movie: The first episode`
...

Chosen Result for `An english movie`: `An english movie` (This is not correct, this is the old 197x movie. But this is an exact match)

The current solution is actually not ideal and a Levenshtein distance
could be a good idea here. This is juste an example.