Build Windows installer on CI jobs
Using all the existing bits we can easily build the Windows installer
for existing versions.

In this commit we:
- Grab the latest version of JF Web (based on the latest unstable)
- Grab the latest version of JF_UX
- Install nsis
- Build an installer as a dedicated artifact

Why; because being able to bundle JF installer as part of the daily
builds increases the chances of finding bugs / getting fixes out. It
also means that people don't have to figure out how to install
ffmpeg...etc. if they want to test an unstable fix.

Limitations
- The installer tags itself as the real version, this needs fixing in
  the upstream installer

- We don't upload the installer as part of the CI to repos.jellyfin,
  this is partially because we don't want users easily finding "a
  released version", which is actually unstable but the installer is
  incorrect.

- Some assumptions have been made about how JF web will name its files.
  Ideally we want a symlink which is repointed on repos.jellyfin which
  points to the latest version. For the moment (ab)use the date naming.