Rename and split existing GH Workflows
Renames and splits up existing GH Workflows into various category types
based on what they interact with.

By using a naming convention is makes it trivial to see which automation
interacts with what. E.g. if there's a problem with PRs we know to look
at the pull-request jobs. Previously, the naming included the catch-all:
automation.yaml

The workflows are unaltered, including triggers and actions. This is
purely a move around:

- ci-* : Runs on PR and commits to specified branches
- project-*, pull-request-* : Relates to those specific components
- release-* : Triggers for a new release

The automation workflow was split down into periodic PR and
project jobs depending on what aspect they touch.