Port CI Tests from Azure pipelines to GH Workflow
Ports CI tests from Azure pipeline.

This commit also drops SonarCloud, which is currently disabled. Since we
can easily add it as nested workflows at a later date if we're going to
enable it. But lets drop it for brevity as it's disabled currently.

Lastly, there's no direct drop in (as far as I can tell) for publishing
code coverage as HTML. Leave this as a todo in a future commit too