fix: drop hevc sdr fallback stream
The client starts playback from a stream with an average bitrate.
In some cases this is HEVC SDR. After a while, the client switches
to the HDR stream, but the server doesn't restart ffmpeg and
the client continues playing the old stream.

Since we have a VideoRange condition, the HEVC SDR fallback stream
is no longer needed.