refactor: use Channels as queueing mechanism for periodic websocket messages
The old implementation used async void as the signature for the event handlers, which might lead to thread pool exhaustion as handlers would fetch the data synchronously despite the async signature on GetDataToSend. Furthermore, on every event, the list of active connections were retrieved inside a `lock` statement, which could also lead to threads being blocked during high traffic scenarios.

Queueing a bool is not great design, but this change only aims to alleviate the thread pool exhaustion.