fix: use the ffmpeg set with cli first
The config file entry is no longer editable with a GUI, and most of our packaging will use the CLI to set the ffmpeg path. This order is easier to work with, as the CLI is easier to modify than the XML config file and can help users fix the ffmpeg path more easily.

Signed-off-by: gnattu <gnattuoc@me.com>