cli: add option to disable network change detection
The NetworkChange API implemented in dotnet is problematic on a lot of platforms. It does not reliably report networking changes and can even cause unhandled exceptions to fail the main thread in extreme cases. Add an option to disable it to allow users to work around the issues caused by this API.

Signed-off-by: gnattu <gnattuoc@me.com>