fix: always show library that is now blocked by tag
The way the whitelist tag currently works is wierd. It requires the entire library to have that tag to be shown, and if the library has the tag, all items inside it inherit that tag, making all items visible. This change aims to make it more sensible by allowing the whitelist to work on a per-item basis, ensuring that the allowed libraries are always shown and the whitelist tag only applies to the items inside the library, not the library itself.

Signed-off-by: gnattu <gnattuoc@me.com>