Restore 255.255.255.255 discovery for Linux
Actually, this address should be removed because we cannot enforce the listening interface with this address at all. However, it seems that all the clients are not yet ready for this change. For now, restore the ability to respond auto discovery message with 255.255.255.255, and perhaps remove it later when clients are ready for the change.

Signed-off-by: gnattu <gnattuoc@me.com>