Allow absolute path mismatch when getting virtual folder info
We have historically required the path stored in the database to exactly match the actual path on the filesystem to retrieve virtual folder information, dating back to the Emby era. However, this approach can cause issues when symbolic links are involved, as the link path may differ from the actual pointed path. To address this, allow matching based solely on the folder name when absolute path checking fails, mitigating such problems.

Signed-off-by: gnattu <gnattuoc@me.com>