Force more compatible transcoding profile for LiveTV
Live TV streams are unlikely to use modern features like HEVC and HDR, and the bitrate is also unlikely to be very high.

This sets a hard limit of 20 Mbps for Live TV transcoding and forces the use of the MPEG-TS container to solve compatibility issues with certain providers/tuners.

Signed-off-by: gnattu <gnattuoc@me.com>