Use SharedStream for LiveTV more restrictively
Only use `SharedHttpStream` for streams that appear to be MPEG-TS streams, and only use a HEAD request when there is no extension. This is the best solution I can think of without making API changes. SharedHttpStream is really uncommon for M3U Tuners where most of them is using HLS, and the ones not using HLS almost uses MPEG-TS streams exclusively. Any other weird Tuners that lies in the mime type and does not provide valid file extension, just treat that stream as normal livestream. If that does not work the only way to make that work, it to add an option to the Tuner config to force the stream type.

Signed-off-by: gnattu <gnattuoc@me.com>