Extract media attachment one by one if the filename appears to be a path
FFmpeg's built-in batch extraction will have trouble writing correct files if the filename appears to be a path. Instead, use the stream index name as the file name in such cases and extract those attachments one by one. This method is slower but works around problematic files that have a path as the filename.

Signed-off-by: gnattu <gnattuoc@me.com>