NextUp query respects Limit
Currently Nextup query does not use the limit defined in the query. It just gets ignored halfway through and defaults to "null". This means that no limit is placed and the sqlite query just pulls everything. This fix uses the query to define a limit X. This should also massively speed up nextup from pulling everything to only pulling X number of items.