Resolve remote images using providers
Currently the resolution of remote images provided by plugins are handled differently in two places;

- If a remote image is acquired internally in Jellyfin it will ask the provider to fetch the image for us.

- If a remote image is asked by a client to be added to an item, it will always be downloaded using a standard HTTP client. I'll also note that the client is sending the image url, provider name, and image type with each request.

This PR changes the behavior of the second scenario to better align with the first scenario, where it will ask the provider to fetch the image if the client tells us which provider to use. Of course, the provider needs to be enabled for the item, otherwise it will result in a `400 Bad Request`. I feel this change is necessary to accommodate when it's not possible for the Jellyfin server itself to directly fetch the images, but a provider can fetch it for the server, all the while the clients can also fetch the images. This can arise when the server and clients use different DNS server and live on different LANs, e.g. a docker network and normal LAN outside the docker network on a different sub-net, both with different DNS servers, where nether have the other DNS server in their DNS stack.